+++ src/Place.h
@@ -16,7 +16,7 @@ public:
     Place(unsigned long id, double dLon, double dLat, double rLon, double rLat, std::pair<int, int> coordinates);
 
     unsigned long getID() const;
-    std::pair<double, double> getCoordinates();
+    std::pair<int, int> getCoordinates();
     std::string toString();
